Interaction Script
What this does-

If a character is close to an interactable object, a Sprite appears
above the object. If the player pushes the interact button (currently
arrow up) while the Sprite is being displayed, text (that describes how
to interact with the object) appears in its place. There is also a bool
called conversation that, if true, activates the objects conversation
script (almost: I do not have the script so I do not know how to call
it, but i wrote an Invoke call for when there is one).